requirements.txt
streamlit>=1.32.0
PyMuPDF>=1.23.0
google-generativeai
python-dotenv